Move Entry into its own file
And lock down the visibility of functions.

This avoids people poking at inner in ways which may be problematic.
